requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-dotenv
sqlalchemy
psycopg2-binary
passlib
passlib[bcrypt]
python-jose[cryptography]
itsdangerous
fastapi_mail
jinja2
pymongo
motor
firebase-admin
